+++ test/bitstring_test.exs
@@ -7,16 +7,10 @@ defmodule IntSetBitstringTest do
   use ExUnitProperties
 
   describe "IntSet.bitstring/1" do
-    property "returns an equal bitstring to that which created it" do
+    property "returns an byte-aligned bitstring" do
       check all int <- positive_integer() do
         bitstr = IntSet.new(int) |> IntSet.bitstring()
-        assert bit_size(bitstr) >= int + 1
-      end
-    end
-
-    property "returns a bitstring big enough to contain all its members" do
-      check all bitstr <- bitstring() do
-        assert IntSet.new(bitstr) |> IntSet.bitstring() == bitstr
+        assert bit_size(bitstr) >= div(int, 8) + 1
       end
     end