+++ test/inspect_test.exs
@@ -10,13 +10,13 @@ defmodule IntSetInspectTest do
     test "inspectable" do
       set = IntSet.new([1, 2, 3])
 
-      assert inspect(set) == "#IntSet<[1, 2, 3]>"
+      assert inspect(set) == "IntSet.new([1, 2, 3])"
     end
 
     test "inspection renders all ints as integers and not characters" do
       set = IntSet.new(7)
 
-      assert inspect(set) == "#IntSet<[7]>"
+      assert inspect(set) == "IntSet.new([7])"
     end
   end
 end